Re-use jwt token instead of logging in every time
This commit ensures we can re-use JWT tokens, and refresh accordingly,
instead of relogging every time, which often causes 429/rate-limits from
checkwatt.

The client of the library should pass in an authinfo object (by
reference) when using the manager. The manager will update any info in
the authinfo object as needed.